Did cleanup for merge and removed new Brain Checker (Logic smth)

+++ userbot/modules/purge.py
@@ -10,25 +10,28 @@ from asyncio import sleep
 from telethon.errors import rpcbaseerrors
 
 from userbot import BOTLOG, BOTLOG_CHATID, CMD_HELP
-from userbot.events import register, errors_handler
+from userbot.events import register
 
 
 @register(outgoing=True, pattern="^.purge$")
-@errors_handler
 async def fastpurger(purg):
     """ For .purge command, purge all messages starting from the reply. """
     chat = await purg.get_input_chat()
     msgs = []
+    itermsg = purg.client.iter_messages(chat, min_id=purg.reply_to_msg_id)
     count = 0
 
-    async for msg in purg.client.iter_messages(chat,
-                                               min_id=purg.reply_to_msg_id):
-        msgs.append(msg)
-        count = count + 1
-        msgs.append(purg.reply_to_msg_id)
-        if len(msgs) == 100:
-            await purg.client.delete_messages(chat, msgs)
-            msgs = []
+    if purg.reply_to_msg_id is not None:
+        async for msg in itermsg:
+            msgs.append(msg)
+            count = count + 1
+            msgs.append(purg.reply_to_msg_id)
+            if len(msgs) == 100:
+                await purg.client.delete_messages(chat, msgs)
+                msgs = []
+    else:
+        await purg.edit("`No message specified.`", )
+        return
 
     if msgs:
         await purg.client.delete_messages(chat, msgs)
@@ -48,7 +51,6 @@ async def fastpurger(purg):
 
 
 @register(outgoing=True, pattern="^.purgeme")
-@errors_handler
 async def purgeme(delme):
     """ For .purgeme, delete x count of your latest message."""
     message = delme.text
@@ -78,7 +80,6 @@ async def purgeme(delme):
 
 
 @register(outgoing=True, pattern="^.del$")
-@errors_handler
 async def delete_it(delme):
     """ For .del command, delete the replied message. """
     msg_src = await delme.get_reply_message()
@@ -96,7 +97,6 @@ async def delete_it(delme):
 
 
 @register(outgoing=True, pattern="^.editme")
-@errors_handler
 async def editer(edit):
     """ For .editme command, edit your last message. """
     message = edit.text
@@ -116,7 +116,6 @@ async def editer(edit):
 
 
 @register(outgoing=True, pattern="^.sd")
-@errors_handler
 async def selfdestruct(destroy):
     """ For .sd command, make seflf-destructable messages. """
     message = destroy.text
@@ -132,26 +131,26 @@ async def selfdestruct(destroy):
 
 
 CMD_HELP.update(
-    {'purge': '.purge'
-     '\nUsage: Purge all messages starting from the reply.'})
+    {'purge': '.purge\n'
+     'Usage: Purge all messages starting from the reply.'})
 
 CMD_HELP.update({
     'purgeme':
-    '.purgeme <x>'
-    '\nUsage: Delete x amount of your latest messages.'
+    '.purgeme <x>\n'
+    'Usage: Delete x amount of your latest messages.'
 })
 
-CMD_HELP.update({"del": ".del" "\nUsage: Delete the message you replied to."})
+CMD_HELP.update({'del': '.del\n' 'Usage: Delete the message you replied to.'})
 
 CMD_HELP.update({
     'editme':
-    ".editme <newmessage>"
-    "\nUsage: Edit the text you replied to with newtext."
+    '.editme <newmessage>\n'
+    'Usage: Edit the text you replied to with newtext.'
 })
 
 CMD_HELP.update({
     'sd':
-    '.sd <x> <message>'
-    "\nUsage: Create a message that self-destructs in x seconds."
-    '\nKeep the seconds under 100 since it puts your bot to sleep.'
+    '.sd <x> <message>\n'
+    "Usage: Create a message that self-destructs in x seconds.\n"
+    'Keep the seconds under 100 since it puts your bot to sleep.'
 })
